requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
dataclasses-json==0.5.14
Deprecated==1.2.14
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
langchain==0.0.268
langsmith==0.0.25
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
packaging==23.1
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
PyGithub==1.59.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.20
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
weaviate-client==3.22.1
wrapt==1.15.0
yarl==1.9.2